Implement task report APIs

+++ routers/official/tasks/employees-tasks.js
@@ -49,6 +49,7 @@ router.get("/", auth, async (req, res) => {
     task.Reports = JSON.parse(task.Reports);
     task.Reports.forEach((report) => {
       report.TaskReportSeens = JSON.parse(report.TaskReportSeens);
+      report.Files = JSON.parse(report.Files);
     });
     task.Supervisors = JSON.parse(task.Supervisors);
     task.Files = JSON.parse(task.Files);
@@ -89,6 +90,7 @@ router.post("/", auth, async (req, res) => {
   result.Reports = JSON.parse(result.Reports);
   result.Reports.forEach((report) => {
     report.TaskReportSeens = JSON.parse(report.TaskReportSeens);
+    report.Files = JSON.parse(report.Files);
   });
   result.Supervisors = JSON.parse(result.Supervisors);
   result.Files = JSON.parse(result.Files);
@@ -117,6 +119,68 @@ router.delete("/:recordID", auth, async (req, res) => {
     dir = `${baseDir}${f.TypeID === 1 ? "task-files" : "task-report-files"}/${
       f.FileName
     }`;
+
+    if (fs.existsSync(dir)) {
+      fs.unlinkSync(dir);
+    }
+  });
+
+  res.send({ Message: result.Message });
+});
+
+router.post("/report", auth, async (req, res) => {
+  const { MemberID } = req.user;
+
+  // second params is <1> : submit report in employees tasks page
+  let result = await selectQuery(
+    `EXEC TaskAPI.SaveTaskReport ${MemberID}, 1, N'${JSON.stringify(req.body)}'`
+  );
+
+  result = result.recordset;
+
+  if (result.Error) return res.status(400).send(result);
+
+  result.forEach((report) => {
+    report.TaskReportSeens = JSON.parse(report.TaskReportSeens);
+    report.Files = JSON.parse(report.Files);
+  });
+
+  res.send(result);
+});
+
+router.post("/report/seen/:taskID", auth, async (req, res) => {
+  const { MemberID } = req.user;
+
+  let result = await selectQuery(
+    `EXEC TaskAPI.MakeReportsSeen ${MemberID}, N'${req.params.taskID}'`
+  );
+
+  result = result.recordset[0];
+
+  if (result.Error) return res.status(400).send(result);
+
+  res.send(result);
+});
+
+router.delete("/report/:recordID", auth, async (req, res) => {
+  const { MemberID } = req.user;
+
+  // second params is <1> : submit report in employees tasks page
+  let result = await selectQuery(
+    `EXEC TaskAPI.DeleteTaskReport ${MemberID}, 1, ${req.params.recordID}`
+  );
+
+  result = result.recordset[0];
+
+  if (result.Error) return res.status(400).send(result);
+
+  result.DeletedFiles = JSON.parse(result.DeletedFiles);
+
+  const baseDir = `./uploaded-files/task-report-files`;
+  let dir = "";
+
+  result.DeletedFiles.forEach((f) => {
+    dir = `${baseDir}/${f.FileName}`;
 
     if (fs.existsSync(dir)) {
       fs.unlinkSync(dir);
